apps/patrimony-back/pyproject.toml
[tool.poetry]
name = "patrimony-back"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "patrimony_back"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.0"
fastapi-crudrouter = "^0.8.6"
pydantic = "^1.10.6"
asyncpg = "^0.27.0"
tortoise-orm = "^0.19.3"
uvicorn = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "patrimony_back:start"